pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "platex-client"
version = "1.0.0"
description = "Windows clipboard assistant that stores image OCR results as LaTeX without overwriting the clipboard."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0.0",
    "pynput>=1.7.6",
    "pystray>=0.19.5",
    "PyQt6>=6.7.0",
    "platformdirs>=4.0.0",
    "PyYAML>=6.0.0",
    "rich>=13.0.0",
]

[project.scripts]
platex-client = "platex_client.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
